test_interactive_legend.py added to test_plots
A simple demo of the interactive legend is added to test_plots. It uses
the same code as the legend example, but now in an interactive form.
When running test_plots.py, the legend labels appear to be cut of the
figure. This still needs to be addressed

In addition there are various changes in the interactive legend plugin
itself in order to handle both the lines and markers in case of Line2D
instances. The used solution is based on the discussion in #194.