Changes as indicated by Aflaxman
Notebook now imports the legend, rather than having the code in the
notebook.

The two superfluous logs to the console have been removed.

The test code on which there was a comment is a straight copy from
test_legend.py with the legend command replaced by the interactive
legend plugin