fix for #188
See discussion under #188 for more details.

In short, the frameon attribute is passed on by mplexporter.
MPLD3renderer makes sure it is being passed to the javascript side.
Axes.js is modified to take into account whether frameon is true or
false. If frameon is false, the background rectangle is not drawn