Complete panning and zooming implementation.
I've changed the general graph components (path, text etc.), to use the new
(standard) D3 zoom implementation. I've commented out the old zooming code,
but have left it in for now in case it's needed.

I've also slightly changed the structure of the elements in the SVG, so that
paths are grouped logically, with the axes etc. being in their own group,
and all the actual graph elements in another.

Currently I'm slowly migrating from the previous structure where everything
was tightly coupled with the axes, and moving towards a more traditional
D3 structure.